refactor(router-core): create RouterCore base class
