+++ packages/router-core/package.json
@@ -60,6 +60,7 @@
   },
   "dependencies": {
     "@tanstack/history": "workspace:*",
-    "@tanstack/store": "^0.7.0"
+    "@tanstack/store": "^0.7.0",
+    "tiny-invariant": "^1.3.3"
   }
 }
