+++ packages/router-core/src/history.ts
@@ -1,0 +1,9 @@
+import type { HistoryLocation } from '@tanstack/history'
+
+declare module '@tanstack/history' {
+  interface HistoryState {
+    __tempLocation?: HistoryLocation
+    __tempKey?: string
+    __hashScrollIntoViewOptions?: boolean | ScrollIntoViewOptions
+  }
+}
